removed firebase, changed version code and number
