+++ Movim/src/main/AndroidManifest.xml
@@ -19,22 +19,6 @@
         android:roundIcon="@mipmap/ic_logo_round"
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
-        <service
-            android:name=".NotificationActivity"
-            android:enabled="true"
-            android:exported="true"></service>
-        <service
-            android:name=".NotificationService"
-            android:enabled="true"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.google.firebase.MESSAGING_EVENT" />
-            </intent-filter>
-        </service>
-
-        <meta-data
-            android:name="firebase_analytics_collection_enabled"
-            android:value="false" />
 
         <activity
             android:name=".ShareActivity"
