+++ Movim/src/main/java/com/movim/movim/MainActivity.java
@@ -1,6 +1,11 @@
 package com.movim.movim;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.HttpURLConnection;
+import java.net.URL;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 
 import android.Manifest;
@@ -8,6 +13,13 @@ import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.AlertDialog.Builder;
+import android.app.Notification;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.app.PendingIntent;
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.IntentFilter;
 import android.content.pm.PackageManager;
 import androidx.annotation.NonNull;
 import androidx.core.app.ActivityCompat;
@@ -15,12 +27,16 @@ import androidx.core.app.ActivityCompat;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.res.Configuration;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.graphics.Color;
 import android.graphics.PixelFormat;
 import android.net.Uri;
 import android.net.http.SslError;
 import android.os.Build;
 import android.os.Bundle;
+
+import androidx.core.app.NotificationCompat;
 import androidx.core.content.ContextCompat;
 import android.text.InputType;
 import android.util.Log;
@@ -43,15 +59,12 @@ import android.widget.Toast;
 
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
-import com.google.firebase.messaging.FirebaseMessaging;
-
-import static com.google.firebase.messaging.Constants.TAG;
 
 @SuppressLint("SetJavaScriptEnabled")
 public class MainActivity extends Activity {
 	private WebView webview;
 	private ProgressBar progressbar;
-	private String firebaseToken = "";
+	private HashMap<String, List<String>> notifs;
 	private static MainActivity instance;
 
 	private ValueCallback<Uri[]> mUploadMessageArray;
@@ -61,6 +74,7 @@ public class MainActivity extends Activity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		setTheme(R.style.SplashTheme);
+		this.notifs = new HashMap<String, List<String>>();
 
 		super.onCreate(savedInstanceState);
 
@@ -77,17 +91,6 @@ public class MainActivity extends Activity {
 		webview.getSettings().setAppCacheEnabled(true);
 		webview.getSettings().setMediaPlaybackRequiresUserGesture(false);
 
-		FirebaseMessaging.getInstance().getToken()
-				.addOnCompleteListener(new OnCompleteListener<String>() {
-					@Override
-					public void onComplete(@NonNull Task<String> task) {
-						if (!task.isSuccessful()) {
-							Log.w(TAG, "Fetching FCM registration token failed", task.getException());
-							return;
-						}
-						firebaseToken = task.getResult();
-					}
-				});
 		
 		if (Build.VERSION.SDK_INT >= 21) {
 			webview.getSettings().setAllowUniversalAccessFromFileURLs(true);
@@ -128,9 +131,6 @@ public class MainActivity extends Activity {
 				if (progress == 100) {
 					progressbar.setVisibility(ProgressBar.GONE);
 
-					if (!firebaseToken.isEmpty()) {
-						webview.loadUrl("javascript:Presence.setFirebaseToken('" + firebaseToken + "')");
-					}
 				}
 			}
 
@@ -228,7 +228,7 @@ public class MainActivity extends Activity {
 	@Override
 	protected void onStop () {
 		super.onStop() ;
-		startService( new Intent( this, NotificationService. class )) ;
+		startService( new Intent( this, NotificationManager. class )) ;
 	}
 
 	private boolean checkAndRequestPermissions() {
@@ -266,7 +266,7 @@ public class MainActivity extends Activity {
 	@Override
 	public void onNewIntent(Intent intent) {
 		if (intent.getAction() != null) {
-			//this.notifs.remove(intent.getAction());
+			this.notifs.remove(intent.getAction());
 			webview.loadUrl(intent.getAction());
 		}
 	}
@@ -310,10 +310,6 @@ public class MainActivity extends Activity {
 		mUploadMessageArray = null;
 	}
 
-	@JavascriptInterface
-	public void clearFirebaseToken() {
-		firebaseToken = "";
-	}
 
 	@JavascriptInterface
 	public void openVisio(String url) {
@@ -328,26 +324,110 @@ public class MainActivity extends Activity {
 		Toast.makeText(this, toast, Toast.LENGTH_SHORT).show();
 	}
 
+
 	@JavascriptInterface
 	public void clearNotifications(String action) {
-		Intent notificationService = new Intent(this, NotificationService.class);
-		notificationService.setAction("clear");
-		notificationService.putExtra("action", action);
-		startService(notificationService);
+		NotificationManager notificationManager = (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
+		notificationManager.cancel(action, 0);
+
+		if (this.notifs.get(action) != null) {
+			this.notifs.remove(action);
+		}
+
+		this.updateNotifications();
 	}
 
 	@JavascriptInterface
 	public void showNotification(String title, String body, String picture, String action) {
-		Intent notificationService = new Intent(this, NotificationService.class);
-		notificationService.setAction("notify");
-		notificationService.putExtra("title", title);
-		notificationService.putExtra("body", body);
-		notificationService.putExtra("picture", picture);
-		notificationService.putExtra("action", action);
-		startService(notificationService);
+		Bitmap pictureBitmap = getBitmapFromURL(picture);
+
+		Intent i = new Intent(this, MainActivity.class);
+		if (action != null) {
+			i.setAction(action);
+			i.addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP | Intent.FLAG_ACTIVITY_CLEAR_TOP);
+		}
+		PendingIntent pi = PendingIntent.getActivity(this, 0, i, PendingIntent.FLAG_UPDATE_CURRENT);
+
+		// The deleteIntent declaration
+		Intent deleteIntent = new Intent(action);
+		PendingIntent pendingDeleteIntent = PendingIntent.getBroadcast(this, 0, deleteIntent,
+				PendingIntent.FLAG_UPDATE_CURRENT);
+		registerReceiver(receiver, new IntentFilter(action));
+
+		// Integer counter;
+		List<String> messages = null;
+
+		// There is already pending notifications
+		if (this.notifs.get(action) != null) {
+			messages = this.notifs.get(action);
+		} else {
+			messages = new ArrayList<String>();
+		}
+
+		messages.add(body);
+
+		if (messages.size() > 5) {
+			messages.remove(0);
+		}
+
+		this.notifs.put(action, messages);
+
+		// We create the inbox
+		NotificationCompat.InboxStyle style = new NotificationCompat.InboxStyle();
+		for (int j = 0; j < messages.size(); j++) {
+			style.addLine(messages.get(j));
+		}
+
+		style.setBigContentTitle(title);
+
+		NotificationManager notificationManager = (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
+
+		String channelId = "channel-movim";
+		String channelName = "Movim";
+		String groupId = "movim";
+		if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
+			NotificationChannel mChannel = new NotificationChannel(channelId, channelName, NotificationManager.IMPORTANCE_DEFAULT);
+			notificationManager.createNotificationChannel(mChannel);
+		}
+
+		Notification notification = new NotificationCompat.Builder(this, channelId)
+				.setSmallIcon(R.drawable.ic_stat_name)
+				.setLargeIcon(pictureBitmap)
+				.setContentTitle(title)
+				.setContentText(body)
+				.setContentIntent(pi)
+				.setDeleteIntent(pendingDeleteIntent).setAutoCancel(true).setColor(Color.parseColor("#3F51B5"))
+				.setLights(Color.parseColor("#3F51B5"), 1000, 5000)
+				.setNumber(messages.size())
+				.setStyle(style)
+				.setGroup(groupId)
+				.build();
+		notificationManager.notify(action, 0, notification);
+
+		if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
+			Notification summaryNotification = new NotificationCompat.Builder(this, channelId)
+					.setSmallIcon(R.drawable.ic_stat_name)
+					.setLargeIcon(BitmapFactory.decodeResource(getResources(), R.drawable.ic_vectorial))
+					.setGroup(groupId)
+					.setGroupSummary(true)
+					.setAutoCancel(true)
+					.build();
+			notificationManager.notify("summary", 0, summaryNotification);
+		}
 	}
 
-	/*private final BroadcastReceiver receiver = new BroadcastReceiver() {
+	protected void updateNotifications() {
+		if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
+			NotificationManager notificationManager = (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
+			int counter = notificationManager.getActiveNotifications().length;
+
+			if (counter <= 1) {
+				notificationManager.cancel("summary", 0);
+			}
+		}
+	}
+
+	private final BroadcastReceiver receiver = new BroadcastReceiver() {
 		@Override
 		public void onReceive(Context context, Intent intent) {
 			if (intent.getAction() != null) {
@@ -355,7 +435,22 @@ public class MainActivity extends Activity {
 			}
 			unregisterReceiver(this);
 		}
-	};*/
+	};
+
+	public static Bitmap getBitmapFromURL(String src) {
+		try {
+			URL url = new URL(src);
+			HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+			connection.setDoInput(true);
+			connection.connect();
+			InputStream input = connection.getInputStream();
+			Bitmap myBitmap = BitmapFactory.decodeStream(input);
+			return myBitmap;
+		} catch (IOException e) {
+			e.printStackTrace();
+			return null;
+		}
+	}
 
 	public boolean onKeyUp(int keyCode, KeyEvent event) {
 		if (keyCode == KeyEvent.KEYCODE_MENU) {
